Use factory pattern to create db instances
This makes the App component control the main DB instance for the app.
If other components down the road want to directly call the database,
they will need to have it passed to them as a prop.

This will also allow us to set up multiple database instances, one for
the database that's synced over IPFS, and, say, another for app UI data
that should not be synced between multiple machines.